Location arg is now optional, added relative coordinates multiplication
The `location` argument is now optional, when `dimension` argument is provided.
When coordinates are relative, they are multiplyed by dimension coordinates scale. This isn't perfect yet, because coordinates are multiplyed AFTER adding relative coordinates and not BEFORE like when using `execute in ..`. I tried fixing it, but I had a lot of problems with it and decided to delay it. Gnoun calculated a function that would help me in getting the correct coordinates so i will probably use it when i will be fixing the error.